python/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["hayata-yamamoto <hayata.yamamoto.work@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
protobuf = "^3.11.3"
grpcio-tools = "^1.29.0"
grpcio-reflection = "^1.29.0"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
yapf = "^0.30.0"
flake8 = "^3.8.1"
lxml = "^4.5.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
